+++ pybind11_geobuf/__main__.py
@@ -1,9 +1,11 @@
 import os
+from typing import Tuple
 
 from loguru import logger
 
 from pybind11_geobuf import rapidjson  # noqa
 from pybind11_geobuf import Decoder, Encoder  # noqa
+from pybind11_geobuf import is_subset_of as is_subset_of_impl  # noqa
 from pybind11_geobuf import normalize_json as normalize_json_impl  # noqa
 from pybind11_geobuf import pbf_decode as pbf_decode_impl  # noqa
 
@@ -90,6 +92,10 @@ def normalize_json(
     sort_keys: bool = True,
     precision: int = -1,
     only_xy: bool = False,
+    denoise_double_0: bool = True,
+    strip_geometry_z_0: bool = True,
+    round_geojson_non_geometry: int = 3,
+    round_geojson_geometry: Tuple[int, int, int] = (8, 8, 3),
 ):
     logger.info(
         f"normalize_json {input_path} ({__filesize(input_path):,} bytes)"
@@ -116,6 +122,10 @@ def normalize_json(
             output_path,
             indent=indent,
             sort_keys=sort_keys,
+            denoise_double_0=denoise_double_0,
+            strip_geometry_z_0=strip_geometry_z_0,
+            round_geojson_non_geometry=round_geojson_non_geometry,
+            round_geojson_geometry=round_geojson_geometry,
         ), f"failed to normalize json to {output_path}"
     logger.info(f"wrote to {output_path} ({__filesize(output_path):,} bytes)")
 
@@ -155,12 +165,6 @@ def round_trip(
     json2pb_use_python: bool = False,
     pb2json_use_python: bool = False,
 ):
-    """
-    _0.json
-    _1.pbf
-    _2.pbf.txt
-    _3.json
-    """
     assert path.endswith((".json", ".geojson")) and os.path.isfile(path)
     path = os.path.abspath(path)
     output_dir = os.path.abspath(output_dir or os.path.dirname(path))
@@ -200,6 +204,10 @@ def round_trip(
     logger.info(f"wrote to {opath}")
 
 
+def is_subset_of(path1: str, path2: str):
+    assert is_subset_of_impl(path1, path2)
+
+
 if __name__ == "__main__":
     import fire
 
@@ -207,6 +215,7 @@ if __name__ == "__main__":
     fire.Fire(
         {
             "geobuf2json": geobuf2json,
+            "is_subset_of": is_subset_of,
             "json2geobuf": json2geobuf,
             "normalize_geobuf": normalize_geobuf,
             "normalize_json": normalize_json,